rr_project/requirements.txt
Django>=4.2.0,<5.0.0
supabase>=1.0.0
python-dotenv>=1.0.0
PyJWT>=2.8.0
django-cors-headers>=4.0.0
psycopg2-binary>=2.9.0  # If using PostgreSQL
requests>=2.31.0